Chad/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
dnspython==1.16.0
eventkit==0.8.8
ib-insync==0.9.66
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
nest-asyncio==1.5.1
numpy==1.21.1
pandas==1.3.1
pandas-ta==0.3.14b0
Pillow==8.3.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
